faiss_service/requirements.txt
numpy
faiss-gpu
pandas
